# use one tri instead of two for each splat