![demo](https://github.com/user-attachments/assets/c6b0ad83-15e8-4d96-a3e8-f9a154eb2dcb)

You only need one tri to render each splat, from my benchmarking this is in the range of 3-7% faster.

The same 99% of the splat is covered but it gives the pixel shader `sqrt(3)` more empty area to render.
The triangle covering the splat is optimal.

I would argue that it is more pleasant for users that number of splats == num of tris.

![image](https://github.com/user-attachments/assets/e7145b7d-a346-46a4-a441-2f786993a74a)