@aras-p 2025-05-22
> You only need one tri to render each splat, from my benchmarking this is in the range of 3-7% faster

On which GPUs do you see that? I suspect which one is faster could be highly dependent on the GPU architecture, and the relative performance of triangle setup vs. ability to discard groups of empty pixels that all take the "discard this pixel" code path.