@wells2kb 2025-05-22
My tests were on a 3080ti laptop gpu.

I could get you exact numbers from my quest 2 tomorrow. But Id be curious to see numbers from other hardware too :)