@aras-p 2025-05-23
On RTX 3080Ti, I see different results depending on resolution. In the "bicycle" scene from the original paper (6.1M splats), rendering at 1200x797 resolution I see a speedup (6.6ms -> 6.3ms), but rendering at 4K resolution I see a slowdown (11.9ms -> 12.5ms)