@wells2kb 2025-05-23
Ah that makes so much sense, I also saw significant slowdown on my quest 2.

I could see a world where this is maybe a config for someone really trying to hit number. But, I don't think that it would be worth while.